Ligero commitment preliminaries (#45)
Various things we will need for Ligero commitments.

Computing a sumcheck proof or generating Ligero linear constraints from
a sumcheck proof requires the Ligero commitment. Computing the
commitment requires the quadratic constraints and the witness. We
refactor the Sumcheck prover so that the witness can be computed
independently (using the new `witness::Witness`) and the constraint
builder so that it can compute quadratic constraints separately from the
linear (using the new `constraints::quadratic_constraints` function).

Introduce a type `LigeroCommitment` wrapping a byte array with some
convenience methods on it.

Set up a test for Ligero commitment that currently fails.