Ligero commitment to witnesses
Construct a Ligero commitment to the witnesses by constructing the
tableau matrix, then building a Merkle tree from columns of that matrix.
The commitment is the root of the tree.

To construct tableau rows, we need the `extend` method from 2.2.1. This
is implemented by evaluating and summing Lagrange basis polynomials.