serialization for Ligero proofs
Serialization of Ligero proofs, which also requires serializing a Merkle
tree node. This follows the implementation in `longfellow-zk`, which
diverges significantly from `draft-google-cfrg-libzk`:

- many fields are fixed length arrays whose length must be computed
  from the tableau layout. This means we can't implement `Codec`, as it
  lacks an affordance for a decoding parameter.
- The serialized proof includes the 32 byte nonces that were hashed into
  the Merkle tree leaves corresponding to the requested columns in the
  proof.
- The lengths of field element runs and the Merkle inclusion proof is a
  4 byte unsigned integer, instead of 3 bytes as elsewhere in the
  protocol.

While we currently cheat and don't implement the subfield for
`GF(2^128)`, we do have to deal with subfields enough to figure out how
many field element runs are in an encoded proof.