Add test vector for MAC circuit in GF(2^128) (#90)
Adds a test vector for a circuit that evaluates a MAC in the binary
field GF(2^128) so that we can exercise that field implementation (and
especially `extend()`) with realistic circuit and input sizes.

This test vector was generated from the C++ implementation. See
`test-vectors/README.md` for discussion. It replaces the extant
longfellow-mac-circuit-902a955fbb22323123aac5b69bdf3442e6ea6f80-1 test
vector, which didn't come with a sumcheck or Ligero proof. This means we
can revisit the `CircuitTestVector` structure to make a number of its
fields non optional, which simplifies a bunch of code.

We also reorganize `test-vectors` a little: the vectors we have so far
are for proving statements from a single circuit. In the near future
we'll be testing proofs across two circuits, and those will likely have
a different format, so we prepare for that eventuality by explicitly
marking our existing test vectors as "one-circuit".

This commit also adds tests against the MAC test vector in existing
modules for circuit evaluation, sumcheck prover, constraint generation,
Ligero prover, Ligero verifier and the end to end test in
`zk_one_circuit`. Most of these tests are currently marked `#[ignore =
"slow test + failing"]`, because they're too slow to run in CI (takes
about 22 seconds to Sumcheck prove) and don't work (probably because
we're not injecting GF(2^128) values into the basis beta when we ought
to).