add `field_element_tests` macro (#106)
The new macro is passed a function which is expected to be generic over
trait `FieldElement` and stamps out a module containing test cases for
each of `FieldP128`, `FieldP256` and `Field2_128`.

I did it this way for two reasons:

- Each field gets its own independent test case that can be run in
  parallel with other tests, instead of a single test case that would
  abort on the first failing field and leave you uncertain about whether
  other fields work properly
- The actual test function is plain ordinary Rust code that is easier to
  read and maintain than a macro expansion

I adopted it in `sumcheck::bind`, only. I considered doing so in
`fields`, too, but that would require unifying the tests
`fields::tests::{extend_x_2, extend_many_sizes}` with the ones in
`fields::field2_128::extend`. That's probably a good idea anyway, but I
don't want to do it in this PR.

The `field2_128` tests are currently tagged `#[ignore]` because the test
cases in `sumcheck::bind` fail. I think this is because they make
assumptions about arithmetic that only hold in large characteristic
fields and/or don't deal with injecting integers into the field.